% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_cube.R
\name{flight_cube}
\alias{flight_cube}
\title{Title}
\usage{
flight_cube(X, pik, EPS = 1e-07)
}
\arguments{
\item{EPS}{}
}
\value{

}
\description{
Title
}
\examples{
N=100
n=30
p=5

z=runif(N)
pik=inclusionprobabilities(z,n)
X=cbind(pik,matrix(rnorm(N*p),c(N,p)))
A=X/pik
flight_cube(X,pik)

}
