% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ineq}
\alias{ineq}
\title{title}
\usage{
ineq(X, pik, B, r, EPS = 1e-07)
}
\arguments{
\item{x}{x}
}
\value{
a vector
}
\description{
description
}
\details{
details
}
\examples{

}
\seealso{
func
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
