% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{landingCube}
\alias{landingCube}
\title{Title}
\usage{
landingCube(X, pikstar, pik, comment = TRUE)
}
\arguments{
\item{comment}{}
}
\value{

}
\description{
Title
}
\examples{
rm(list = ls())
EPS=0.0000001
N <- 1000
n <- 300
p <- 1
q <- 7
z <- runif(N)
#z <- rep(1,N)
pik <- inclusionprobabilities(z,n)
X <- cbind(pik,matrix(rnorm(N*p),c(N,p)))
A <- X/pik
Z <- cbind(matrix(rbinom(N*q,1,1/2),c(N,q)))
B <- cbind(Z,-Z)
r <- c(ceiling(pik\%*\%B))
r[abs(pik\%*\%B-round(pik\%*\%B))<EPS] <- round(pik\%*\%B)[abs(pik\%*\%B-round(pik\%*\%B))<EPS]

Astar <- cbind(X,B*pik)
pikstar <- ineq(Astar,pik,B,r,EPS=0.000001)
landingCube(X,pikstar,pik)
}
